+++ src/main/java/com/global/api/entities/gpApi/GpApiRequest.java
@@ -44,4 +44,6 @@ public class GpApiRequest extends Request {
     public static final String FILE_PROCESSING = "/files";
 
     public static final String PAYERS_ENDPOINT = "/payers";
+
+    public static final String INSTALLMENT_ENDPOINT = "/installments";
 }