+++ src/main/java/com/global/api/serviceConfigs/GpApiConfig.java
@@ -113,6 +113,8 @@ public class GpApiConfig extends GatewayConfig {
         services.setPayFacProvider(gpApiConnector);
         services.setSecure3dProvider(Secure3dVersion.ONE, gpApiConnector);
         services.setSecure3dProvider(Secure3dVersion.TWO, gpApiConnector);
+
+        services.setInstallmentService(gpApiConnector);
     }
 
     @Override